requirements.txt
# Python 3.10+
# Core framework
langchain>=0.2.6
langchain-core>=0.2.6
langchain-community>=0.2.6

# HTML parsing
beautifulsoup4>=4.12.2

# AWS Bedrock (LLM re-ranking)
langchain-aws>=0.1.6
boto3>=1.34.0

# Environment variable loading
python-dotenv>=1.0.1

# Async HTTP client (used by async loaders)
aiohttp>=3.9.5
